+++ relayer-cli/src/commands/listen.rs
@@ -1,14 +1,16 @@
 use std::{ops::Deref, sync::Arc, thread};
 
 use abscissa_core::{application::fatal_error, error::BoxError, Command, Options, Runnable};
-use crossbeam_channel as channel;
 use itertools::Itertools;
 use tokio::runtime::Runtime as TokioRuntime;
 
 use tendermint_rpc::query::{EventType, Query};
 
 use ibc::ics24_host::identifier::ChainId;
-use ibc_relayer::{config::ChainConfig, event::monitor::*};
+use ibc_relayer::{
+    config::ChainConfig,
+    event::monitor::{EventMonitor, EventReceiver},
+};
 
 use crate::prelude::*;
 
@@ -73,7 +75,7 @@ fn subscribe(
     chain_config: &ChainConfig,
     queries: Vec<Query>,
     rt: Arc<TokioRuntime>,
-) -> Result<(EventMonitor, channel::Receiver<EventBatch>), BoxError> {
+) -> Result<(EventMonitor, EventReceiver), BoxError> {
     let (mut event_monitor, rx) = EventMonitor::new(
         chain_config.id.clone(),
         chain_config.websocket_addr.clone(),