+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Jongwhan Lee (@leejw51crypto) ([#878]).
 - [ibc-relayer]
   - Change the default for client creation to allow governance recovery in case of expiration or misbehaviour ([#785])
   - Use a single supervisor to subscribe to all configured chains ([#862])
+  - The relayer is now more resilient to a node not being up or going down, and will attempt to reconnect ([#871])
 
 ### BUG FIXES
 
@@ -58,6 +59,7 @@ Jongwhan Lee (@leejw51crypto) ([#878]).
 [#862]: https://github.com/informalsystems/ibc-rs/issues/862
 [#863]: https://github.com/informalsystems/ibc-rs/issues/863
 [#869]: https://github.com/informalsystems/ibc-rs/issues/869
+[#871]: https://github.com/informalsystems/ibc-rs/issues/871
 [#878]: https://github.com/informalsystems/ibc-rs/issues/878
 [#909]: https://github.com/informalsystems/ibc-rs/issues/909
 [#873]: https://github.com/informalsystems/ibc-rs/issues/873